backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
Pillow==11.1.0
django-ckeditor==6.7.0
gunicorn==21.2.0
